+++ src/common/output/debug.h
@@ -37,9 +37,11 @@ int debug_channel	__attribute__((weak)) = 2;
 
 #ifdef SHOW_DEBUGS
 #define debug(...) \
+				{ \
         dprintf(debug_channel, "%s: ", __FUNCTION__); \
         dprintf(debug_channel, __VA_ARGS__); \
-        dprintf(debug_channel, "\n");
+        dprintf(debug_channel, "\n"); \
+				}
 #else
 #define debug(...)
 #endif